requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
greenlet==3.1.1
idna==3.10
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
psycopg2==2.9.10
PyMySQL==1.1.1
pytest==8.3.3
requests==2.32.3
typing_extensions==4.12.2
urllib3==2.2.3
